// packages/core/src/render3/interfaces/injector.d.ts
export declare const enum NodeInjectorOffset {
    TNODE = 8,
    PARENT = 8,
    BLOOM_SIZE = 8,
    SIZE = 9
}
/**
 * Represents a relative location of parent injector.
 *
 * The interfaces encodes number of parents `LView`s to traverse and index in the `LView`
 * pointing to the parent injector.
 */
export type RelativeInjectorLocation = number & {
    __brand__: 'RelativeInjectorLocationFlags';
};
export declare const enum RelativeInjectorLocationFlags {
    InjectorIndexMask = 32767,
    ViewOffsetShift = 16,
    NO_PARENT = -1
}
export declare const NO_PARENT_INJECTOR: RelativeInjectorLocation;